Add files via upload
To do list  application

This change consolidates the front end part of the project tasks. The application allows you to create, add and delete the list items on your to do list. You can even update your list by marking your list items done. The documentation includes the purpose and the wire-frames of the application.
The on-boarding screen is just a temporary fix to make things work; please avoid it